+++ rsconnect/main.py
@@ -145,12 +145,12 @@ def server_args(func):
     return wrapper
 
 
-def rstudio_args(func):
+def cloud_shinyapps_args(func):
     @click.option(
         "--account",
         "-A",
         envvar=["SHINYAPPS_ACCOUNT"],
-        help="The shinyapps.io account name.",
+        help="The shinyapps.io/Posit Cloud account name.",
     )
     @click.option(
         "--token",
@@ -412,7 +412,7 @@ def bootstrap(
     help="The path to trusted TLS CA certificates.",
 )
 @click.option("--verbose", "-v", is_flag=True, help="Print detailed messages.")
-@rstudio_args
+@cloud_shinyapps_args
 @click.pass_context
 def add(ctx, name, server, api_key, insecure, cacert, account, token, secret, verbose):
 
@@ -961,7 +961,7 @@ def deploy_voila(
 )
 @server_args
 @content_args
-@rstudio_args
+@cloud_shinyapps_args
 @click.argument("file", type=click.Path(exists=True, dir_okay=True, file_okay=True))
 @cli_exception_handler
 def deploy_manifest(
@@ -1133,12 +1133,13 @@ def deploy_quarto(
 # noinspection SpellCheckingInspection,DuplicatedCode
 @deploy.command(
     name="html",
-    short_help="Deploy html content to Posit Connect.",
-    help=("Deploy an html file, or directory of html files with entrypoint, to Posit Connect."),
+    short_help="Deploy html content to Posit Connect or Posit Cloud.",
+    help=("Deploy an html file, or directory of html files with entrypoint, to Posit Connect or Posit Cloud."),
     no_args_is_help=True,
 )
 @server_args
 @content_args
+@cloud_shinyapps_args
 @click.option(
     "--entrypoint",
     "-e",
@@ -1177,6 +1178,9 @@ def deploy_html(
     api_key: str = None,
     insecure: bool = False,
     cacert: typing.IO = None,
+    account: str = None,
+    token: str = None,
+    secret: str = None,
 ):
     kwargs = locals()
     ce = None
@@ -1218,7 +1222,7 @@ def generate_deploy_python(app_mode, alias, min_version):
     )
     @server_args
     @content_args
-    @rstudio_args
+    @cloud_shinyapps_args
     @click.option(
         "--entrypoint",
         "-e",