+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - getdefaultlocale deprecated
 
+## Unreleased
+
+### Fixed
+
+- cacert read error when adding/updating a server.
+
 ## [1.16.0] - 2023-03-27
 
 ### Added