+++ CHANGELOG.md
@@ -3,10 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## Unreleased
+## [1.17.0] - 2023-XX-XX
 
 ### Added
+- `deploy html` and `deploy manifest` now support deployment to Posit Cloud.
+
 - Added `system caches list` and `system caches delete` commands which allow administrators to enumerate and delete R and Python runtime caches from Connect servers [#384](https://github.com/rstudio/rsconnect-python/pull/384). Read more about the feature in our [docs](https://docs.posit.co/connect/admin/server-management/runtime-caches/).
+
+### Changed
+- Cloud deployments accept the content id instead of application id in the --app-id field.
+- The `app_id` field in application store files also stores the content id instead of the application id.
+- Application store files include a `version` field, set to 1 for this release.
 
 ### Fixed
 - cacert read error when adding/updating a server [#403](https://github.com/rstudio/rsconnect-python/issues/403).